planeBoundary default constructor, EOF in MPM file, MaterialType default constructor, warning in getNumNodes function, warning in getInsidePolygon function, warning in getBodyList() function, warning in setPrescribedPorePressureBox function, warning in getResultSteps function, removing unused operating system functions

+++ inc/Mesh/Mesh.h
@@ -92,7 +92,7 @@ public:
 
     /// \brief Return total nodes in mesh
     /// \return Total nodes in the mesh
-    inline int getNumNodes() const { return this->gridNodes.size(); }
+    inline unsigned int getNumNodes() const { return (unsigned int) this->gridNodes.size(); }
     
     /// \brief Return nodes in mesh
     /// \return A pointer to the mesh nodes
